// generated/prisma/enums.ts
export const NotificationPriority = {
  LOW: 'LOW',
  NORMAL: 'NORMAL',
  HIGH: 'HIGH',
  URGENT: 'URGENT'
} as const

export type NotificationPriority = (typeof NotificationPriority)[keyof typeof NotificationPriority]


export const NotificationChannel = {
  IN_APP: 'IN_APP',
  EMAIL: 'EMAIL',
  SMS: 'SMS',
  WHATSAPP: 'WHATSAPP'
} as const

export type NotificationChannel = (typeof NotificationChannel)[keyof typeof NotificationChannel]


export const NotificationCategory = {
  GENERAL: 'GENERAL',
  STOCK_ALERT: 'STOCK_ALERT',
  CASH_SESSION: 'CASH_SESSION',
  SALE_COMPLETED: 'SALE_COMPLETED',
  PURCHASE_ORDER: 'PURCHASE_ORDER',
  SYSTEM: 'SYSTEM',
  TIME_CARD: 'TIME_CARD',
  EXPENSE_ALERT: 'EXPENSE_ALERT'
} as const

export type NotificationCategory = (typeof NotificationCategory)[keyof typeof NotificationCategory]